+++ tests/conftest.py
@@ -51,3 +51,8 @@ def text(request):
 @pytest.fixture(params=platform)
 def path(request):
     return mimesis.Path(request.param)
+
+
+@pytest.fixture(params=locales)
+def transport(request):
+    return mimesis.Transport(request.param)